# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API con Express y TypeScript",
    "version": "1.0.0",
    "description": "Documentación generada con swagger-jsdoc"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/talla": {
      "get": {
        "summary": "Obtiene todas las tallas",
        "tags": [
          "Tallas"
        ],
        "responses": {
          "200": {
            "description": "Lista de tallas"
          }
        }
      },
      "post": {
        "summary": "Crea una nueva talla (requiere autenticación)",
        "tags": [
          "Tallas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  }
                },
                "required": [
                  "nombre"
                ]
              },
              "example": {
                "talla": "XXL"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Talle creado"
          },
          "500": {
            "description": "Error al crear el talle"
          }
        }
      }
    },
    "/talla/{id}": {
      "get": {
        "summary": "Obtiene una talla por ID",
        "tags": [
          "Tallas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Talla encontrada"
          },
          "404": {
            "description": "Talla no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualiza una talla por ID (requiere autenticación)",
        "tags": [
          "Tallas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "activo": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "talla": "XL",
                "activo": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Talle actualizado"
          },
          "404": {
            "description": "Talle no encontrado"
          },
          "500": {
            "description": "Error al actualizar el talle"
          }
        }
      },
      "patch": {
        "summary": "Activa o desactiva una talla (requiere autenticación)",
        "tags": [
          "Tallas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activo": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "404": {
            "description": "Talle no encontrado"
          },
          "500": {
            "description": "Error al patchear el talle"
          }
        }
      }
    },
...
  },
...
}